requirements.txt
amqp==2.6.1
billiard==3.6.4.0
celery==4.4.0
cffi==1.15.1
configparser==4.0.2
contextlib2==0.6.0.post1
Django==1.11
django-tastypie==0.14.0
enum34==1.1.10
idna==2.10
importlib-metadata==2.1.3
ipaddress==1.0.23
kombu==4.6.11
mysqlclient==1.4.6
pathlib2==2.3.7.post1
pyasn1==0.5.1
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-mimeparse==1.6.0
pytz==2024.1
scandir==1.10.0
six==1.16.0
typing==3.10.0.0
vine==1.3.0
zipp==1.2.0
